[Vue]The page is redirected to the home page of the website after clicking the "Change associated content" button in the Experience Editor
