+++ packages/sitecore-jss/src/utils/editing.ts
@@ -80,6 +80,45 @@ export const resetEditorChromes = (): void => {
 };
 
 /**
+ * @description in Experience Editor, anchor tags
+ * with both onclick and href attributes will use the href, blocking the onclick from firing.
+ * This function makes it so the anchor tags function as intended in the sample when using Experience Editor
+ *
+ * The Mutation Observer API is used to observe changes to the body, then select all elements with href="#" and an onclick,
+ * and replaces the # value with javascript:void(0); which prevents the anchor tag from blocking the onclick event handler.
+ * @see Mutation Observer API: https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver/MutationObserver
+ */
+export const handleEditorAnchors = () => {
+  // The sample gives the href attribute priority over the onclick attribute if both are present, so we must replace
+  // the href attribute to avoid overriding the onclick in Experience Editor
+  if (!window || !ExperienceEditor.isActive()) {
+    return;
+  }
+  const targetNode = document.querySelector('body');
+  const callback = (mutationList: MutationRecord[]) => {
+    mutationList.forEach((mutation: MutationRecord) => {
+      const btns: NodeListOf<HTMLAnchorElement> = document.querySelectorAll(
+        '.scChromeDropDown > a[href="#"], a[onclick]'
+      );
+      if (mutation.type === 'childList') {
+        btns.forEach((link: HTMLAnchorElement) => {
+          link.href = 'javascript:void(0);';
+        });
+      }
+      return;
+    });
+  };
+  const observer: MutationObserver = new MutationObserver(callback);
+  const observerOptions = {
+    childList: true,
+    subtree: true,
+  };
+  if (targetNode) {
+    observer.observe(targetNode, observerOptions);
+  }
+};
+
+/**
  * Determines whether the current execution context is within the Sitecore Experience Editor
  * @deprecated Will be removed in a future release. Please use isEditorActive instead.
  * @returns true if executing within the Sitecore Experience Editor
