+++ packages/sitecore-jss/src/utils/index.ts
@@ -7,4 +7,5 @@ export {
   resetEditorChromes,
   isExperienceEditorActive,
   resetExperienceEditorChromes,
+  handleEditorAnchors,
 } from './editing';
